Improve the git repository fetch handling
This PR improves support for cloning git repositories, with the
previous setup you could get in a situation where you would fail to
checkout an existing remote branch because it did not exist locally.

Now in the case where the repo is already on disk, we start by fetching
so that we can pull or checkout any commit or tag.